Extract announcement operation to AnnouncementUtils
